requirements.txt
biopython
h5py
hydra-core
jupyter
maskedtensor
matplotlib
numpy
omegaconf
pandas
polars
protobuf==3.20.*
pyarrow
rich
scikit-learn
seaborn
tensorboard
torchmetrics
tqdm